feat: Learn about the State Lifting Mechanism in React
Understand Why we need state lifting mechanism in React. What is it and how to implement it.
