+++ src/playground/demo.js
@@ -1,0 +1,20 @@
+function GrandParent() {
+	// App
+
+	Child1();
+	Child2();
+}
+
+function Child1() {
+	const noteTitle = "title";
+	// return noteTitle
+	// NoteTakingForm
+}
+
+function Child2() {
+	const notes = [];
+	const editMode = false;
+	// NoteList
+}
+
+GrandParent();
